LayoutTests/fast/page-color-sampling/page-color-sampling-skips-border-top.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true obscuredInset.top=100 ] -->
<html>
<head>
<meta charset="utf-8">
<meta name="viewport" content="width=device-width, initial-scale=1">
<style>
body {
    font-family: system-ui;
}

header {
    position: fixed;
    top: 0;
    left: 0;
    width: 100%;
    height: 100px;
    border-top: rgb(231, 112, 13) solid 3px;
    border-bottom: lightgray solid 2px;
    box-sizing: border-box;
    background: white;
}

.tall {
    width: 10px;
    height: 2000px;
}

#description {
    margin-top: 100px;
}
</style>
<script src="../../resources/js-test.js"></script>
<script src="../../resources/ui-helper.js"></script>
<script>
jsTestIsAsync = true;

async function sampledTopColor() {
    return (await UIHelper.fixedContainerEdgeColors()).top;
}

addEventListener("load", async () => {
    let header = document.querySelector("header");
    description("To manually test, scroll this page down; the top sampled color should remain solid white");

    await UIHelper.ensurePresentationUpdate();
    initialTopColor = await sampledTopColor();
    shouldBeEqualToString("initialTopColor", "rgb(255, 255, 255)");

    while (pageYOffset < 500) {
        scrollBy(0, 10);
        const topColor = await sampledTopColor();
        if (topColor !== "rgb(255, 255, 255)")
            testFailed(`Unexpected color sample: ${topColor} at offset ${pageYOffset}`);
    }

    header.style.borderTop = "rgb(238, 238, 238) solid 80px";
    await UIHelper.waitForFixedContainerEdgeColors({
        top: "rgb(238, 238, 238)",
        left: null,
        right: null,
        bottom: null
    });
    testPassed("Top color became rgb(238, 238, 238)");
    finishJSTest();
});
</script>
</head>
<body>
    <header></header>
    <p id="description"></p>
    <div class="tall"></div>
</body>
</html>
